all functions migrated and tested (need to test binance US)
