+++ models/exchange/binance/__init__.py
@@ -1,1 +1,1 @@
-from .api2 import *
+from .api import *
