+++ models/exchange/binance/api.py
@@ -1,24 +1,31 @@
-from models.exchange.coinbase_pro.api import FREQUENCY_EQUIVALENTS, SUPPORTED_GRANULARITY
-import math
-import re
-import numpy as np
-import pandas as pd
-import sys
-from datetime import datetime, timedelta
-from binance.client import Client
-from time import sleep
-from models.helper.LogHelper import Logger
+"""Remotely control your Binance account via their API : https://binance-docs.github.io/apidocs/spot/en"""
 
+import re
+import json
+import hmac
+import hashlib
+import time
+import requests
+import base64
+import sys
+import math
+import pandas as pd
+import numpy as np
+from numpy import floor
+from datetime import datetime, timedelta
+from requests.auth import AuthBase
+from requests import Request, Session
+from models.helper.LogHelper import Logger
+from urllib.parse import urlencode
 
 DEFAULT_MAKER_FEE_RATE = 0.0015 # added 0.0005 to allow for price movements
 DEFAULT_TAKER_FEE_RATE = 0.0015 # added 0.0005 to allow for price movements
 DEFAULT_TRADE_FEE_RATE = 0.0015 # added 0.0005 to allow for price movements
-DEFAULT_GRANULARITY="1h"
-SUPPORTED_GRANULARITY = ['1m', '5m', '15m', '1h', '6h', '1d']
-MULTIPLIER_EQUIVALENTS = [1, 5, 15, 60, 360, 1440]
-FREQUENCY_EQUIVALENTS = ["T", "5T", "15T", "H", "6H", "D"]
-DEFAULT_MARKET = "BTCGBP"
-
+DEFAULT_GRANULARITY ='1h'
+SUPPORTED_GRANULARITY = [ '1m', '5m', '15m', '1h', '6h', '1d' ]
+MULTIPLIER_EQUIVALENTS = [ 1, 5, 15, 60, 360, 1440 ]
+FREQUENCY_EQUIVALENTS = [ 'T', '5T', '15T', 'H', '6H', 'D']
+DEFAULT_MARKET = 'BTCGBP'
 
 class AuthAPIBase():
     def _isMarketValid(self, market: str) -> bool:
@@ -27,9 +34,8 @@ class AuthAPIBase():
             return True
         return False
 
-
 class AuthAPI(AuthAPIBase):
-    def __init__(self, api_key: str='', api_secret: str='', api_url: str='https://api.binance.com') -> None:
+    def __init__(self, api_key: str='', api_secret: str='', api_url: str='https://api.binance.com', order_history: list=[]) -> None:
         """Binance API object model
     
         Parameters
@@ -38,20 +44,19 @@ class AuthAPI(AuthAPIBase):
             Your Binance account portfolio API key
         api_secret : str
             Your Binance account portfolio API secret
+        api_url
+            Binance API URL
         """
-    
+
         # options
         self.debug = False
         self.die_on_api_error = False
 
         valid_urls = [
-            'https://api.binance.com/',
-            'https://api.binance.us/',
-            'https://testnet.binance.vision/api/'
+            'https://api.binance.com',
+            'https://api.binance.us',
+            'https://testnet.binance.vision'
         ]
-
-        if len(api_url) > 1 and api_url[-1] != '/':
-            api_url = api_url + '/'
 
         # validate Binance API
         if api_url not in valid_urls:
@@ -67,29 +72,12 @@ class AuthAPI(AuthAPIBase):
         if not p.match(api_secret):
             self.handle_init_error('Binance API secret is invalid')
 
-        self.mode = 'live' # TODO: check if this needs to be set here
-        self.api_url = api_url
-        self.api_key = api_key
-        self.api_secret = api_secret
-        self.order_attempts = 0
+        self._api_key = api_key
+        self._api_secret = api_secret
+        self._api_url = api_url
 
-        for i in range(10):
-            try:
-                sys.tracebacklimit = 0
-                if 'api.binance.us' in api_url:
-                    self.client = Client(self.api_key, self.api_secret, { 'verify': False, 'timeout': 20 }, tld='us')
-                else:
-                    self.client = Client(self.api_key, self.api_secret, { 'verify': False, 'timeout': 20 })
-                break
-            except Exception as e:
-                if i == 9:
-                    raise SystemExit("Can not create instance of AuthAPI client.")
-                Logger.error('Exception: ' + str(e)) 
-                Logger.error('Error on creating instance of AuthAPI Client. Trying again... Attempt: ' + str(i))
-                sleep(0.1)
-
-        sys.tracebacklimit = 1
-
+        # order history
+        self.order_history = order_history
 
     def handle_init_error(self, err: str) -> None:
         if self.debug:
@@ -97,68 +85,114 @@ class AuthAPI(AuthAPIBase):
         else:
             raise SystemExit(err)
 
+    def _dispatch_request(self, method: str):
+        session = Session()
+        session.headers.update({
+            'Content-Type': 'application/json; charset=utf-8',
+            'X-MBX-APIKEY': self._api_key
+        })
+        return {
+            'GET': session.get,
+            'DELETE': session.delete,
+            'PUT': session.put,
+            'POST': session.post,
+        }.get(method, 'GET')
 
-    def getClient(self) -> Client:
-        return self.client
+    def createHash(self, uri: str=''):
+        return hmac.new(self._api_secret.encode('utf-8'), uri.encode('utf-8'), hashlib.sha256).hexdigest()
 
+    def getTimestamp(self):
+        return int(time.time() * 1000)
 
-    def getAccount(self):
-        """Retrieves a specific account"""
-        account = self.client.get_account()
-        if 'balances' in account:
-            df = pd.DataFrame(account['balances'])
-            df = df[(df['free'] != '0.00000000') & (df['free'] != '0.00')]
-            df['free'] = df['free'].astype(float)
-            df['locked'] = df['locked'].astype(float)
-            df['balance'] = df['free'] - df['locked']
-            df.columns = ['currency', 'available', 'hold', 'balance']
-            df = df[['currency', 'balance', 'hold', 'available']]
-            df = df.reset_index(drop=True)
-            return df
+    def getAccounts(self) -> pd.DataFrame:
+        """Retrieves your list of accounts"""
+
+        # GET /api/v3/account
+        resp = self.authAPI('GET', '/api/v3/account')
+
+        if 'balances' in resp:
+            balances = resp['balances']
+
+            if isinstance(balances, list):
+                df = pd.DataFrame.from_dict(balances)
+            else: 
+                df = pd.DataFrame(balances, index=[0])
         else:
-            return 0.0
+            return pd.DataFrame()
+  
+        if len(df) == 0:
+            return pd.DataFrame()
 
+        # exclude accounts that are locked
+        df = df[df.locked != 0.0]
+        df['locked'] = df['locked'].astype(bool)
+
+        # reset the dataframe index to start from 0
+        df = df.reset_index()
+
+        df['id'] = df['index']
+        df['hold'] = 0.0
+        df['profile_id'] = None
+        df['available'] = df['free']
+
+        df['id'] = df['id'].astype(object)
+        df['hold'] = df['hold'].astype(object)
+
+        # exclude accounts with a nil balance
+        df = df[df.available != '0.00000000']
+        df = df[df.available != '0.00']
+
+        # rename columns
+        df.columns = ['index', 'currency', 'balance', 'trading_enabled', 'id', 'hold', 'profile_id', 'available']
+
+        return df[['index', 'id', 'currency', 'balance', 'hold', 'available', 'profile_id', 'trading_enabled' ]]
+
+    def getAccount(self, account: int) -> pd.DataFrame:
+        """Retrieves a specific account"""
+
+        # validates the account is syntactically correct
+        if not isinstance(account, int):
+            self.handle_init_error('Binance account is invalid')
+    
+        df = self.getAccounts()
+        return df[df.id == account]
 
     def getFees(self, market: str='') -> pd.DataFrame:
-        if market != '':
-            resp = self.client.get_trade_fee(symbol=market)
-            if 'tradeFee' in resp and len(resp['tradeFee']) > 0:
-                df = pd.DataFrame(resp['tradeFee'][0], index=[0])
-                df['usd_volume'] = None
-                df.columns = [ 'maker_fee_rate', 'market', 'taker_fee_rate', 'usd_volume' ]
-                return df[[ 'maker_fee_rate', 'taker_fee_rate', 'usd_volume', 'market' ]]
-            return pd.DataFrame(columns=[ 'maker_fee_rate', 'taker_fee_rate', 'market' ])
-        else:
-            resp = self.client.get_trade_fee()
-            if 'tradeFee' in resp:
-                df = pd.DataFrame(resp['tradeFee'])
-                df['usd_volume'] = None
-                df.columns = [ 'maker_fee_rate', 'market', 'taker_fee_rate', 'usd_volume' ]
-                return df[[ 'maker_fee_rate', 'taker_fee_rate', 'usd_volume', 'market' ]]
-            return pd.DataFrame(columns=[ 'maker_fee_rate', 'taker_fee_rate', 'market' ])
+        """Retrieves a account fees"""
 
+        # GET /api/v3/account
+        resp = self.authAPI('GET', '/api/v3/account')
+
+        if 'makerCommission' in resp and 'takerCommission' in resp:
+            maker_fee_rate = resp['makerCommission'] / 10000
+            taker_fee_rate = resp['takerCommission'] / 10000
+        else:
+            maker_fee_rate = 0.001
+            taker_fee_rate = 0.001
+      
+        return pd.DataFrame([{ 'maker_fee_rate': maker_fee_rate, 'taker_fee_rate': taker_fee_rate, 'usd_volume': 0, 'market': '' }])
 
     def getMakerFee(self, market: str='') -> float:
-        if market == '':
-            fees = self.getFees()
-        else:
+        """Retrieves the maker fee"""
+        
+        if len(market):
             fees = self.getFees(market)
+        else:
+            fees = self.getFees()
         
         if len(fees) == 0 or 'maker_fee_rate' not in fees:
             Logger.error(f"error: 'maker_fee_rate' not in fees (using {DEFAULT_MAKER_FEE_RATE} as a fallback)")
             return DEFAULT_MAKER_FEE_RATE
 
-        if market == '':
-            return fees
-        else:
-            return float(fees['maker_fee_rate'].to_string(index=False).strip())
-
-
+        return float(fees['maker_fee_rate'].to_string(index=False).strip())
+   
     def getTakerFee(self, market: str='') -> float:
-        if market == '':
-            return DEFAULT_TAKER_FEE_RATE
-        else:
+        """Retrieves the taker fee"""
+
+        if len(market) != None:
             fees = self.getFees(market)
+        else:
+            fees = self.getFees()
 
         if len(fees) == 0 or 'taker_fee_rate' not in fees:
             Logger.error(f"error: 'taker_fee_rate' not in fees (using {DEFAULT_TAKER_FEE_RATE} as a fallback)")
@@ -166,22 +200,45 @@ class AuthAPI(AuthAPIBase):
 
         return float(fees['taker_fee_rate'].to_string(index=False).strip())
 
+    def getUSDVolume(self) -> float:
+        """Retrieves the USD volume"""
 
-    def __convertStatus(self, val: str) -> str:
-        if val == 'filled':
-            return 'done'
+        fees = self.getFees()
+        return float(fees['usd_volume'].to_string(index=False).strip())
+
+    def getMarkets(self) -> list:
+        """Retrieves a list of markets on the exchange"""
+
+        # GET /api/v3/allOrders
+        resp = self.authAPI('GET', '/api/v3/exchangeInfo')
+
+        if 'symbols' in resp:
+            if isinstance(resp['symbols'], list):
+                df = pd.DataFrame.from_dict(resp['symbols'])
+            else: 
+                df = pd.DataFrame(resp['symbols'], index=[0])
         else:
-            return val
+            df = pd.DataFrame()
 
+        return df[df['isSpotTradingAllowed'] == True][['symbol']].squeeze().tolist()
 
-    def getOrders(self, market: str='', action: str='', status: str='all') -> pd.DataFrame:
+    def getOrders(self, market: str='', action: str='', status: str='all', order_history: list=[]) -> pd.DataFrame:
         """Retrieves your list of orders with optional filtering"""
 
         # if market provided
+        markets = None
         if market != '':
             # validates the market is syntactically correct
             if not self._isMarketValid(market):
                 raise ValueError('Binance market is invalid.')
+        else:
+            if len(order_history) > 0:
+                full_scan = False
+                self.order_history = order_history
+                markets = self.order_history
+            else:
+                full_scan = True
+                markets = self.getMarkets()
 
         # if action provided
         if action != '':
@@ -189,64 +246,153 @@ class AuthAPI(AuthAPIBase):
             if not action in ['buy', 'sell']:
                 raise ValueError('Invalid order action.')
 
-        # validates status is either open, pending, done, active, or all
-        if not status in ['open', 'pending', 'done', 'active', 'all']:
+        # validates status is either open, canceled, pending, done, active, or all
+        if not status in ['open', 'canceled', 'pending', 'done', 'active', 'all']:
             raise ValueError('Invalid order status.')
 
-        def attempt_orders():
-            resp = self.client.get_all_orders(symbol=market)
-            if len(resp) > 0:
-                self.order_attempts = 0
-                return pd.DataFrame(resp)
-            else:
-                if self.order_attempts < 3:
-                    self.order_attempts += 1
-                    Logger.debug(f'No previous orders found. Retrying: attempt {self.order_attempts}/3')
-                    return attempt_orders()
-                else:
-                    Logger.debug(f'Exhausted attempts to retrieve orders')
-                    return pd.DataFrame()
+        if markets is not None:
+            df = pd.DataFrame()
+            for market in markets:
+                if full_scan is True:
+                    print (f'scanning {market} order history.')
 
-        df = attempt_orders()
+                # GET /api/v3/allOrders
+                resp = self.authAPI('GET', '/api/v3/allOrders', { 'symbol': market })
+
+                if full_scan is True:
+                    time.sleep(0.25)
+
+                if isinstance(resp, list):
+                    df_tmp = pd.DataFrame.from_dict(resp)
+                else: 
+                    df_tmp = pd.DataFrame(resp, index=[0])
+
+                if full_scan is True and len(df_tmp) > 0:
+                    self.order_history.append(market)
+                
+                if len(df_tmp) > 0:
+                    df = pd.concat([df, df_tmp])
+
+            if full_scan is True:
+                print ('add to order history to prevent full scan:', self.order_history)
+        else:
+            # GET /api/v3/allOrders
+            resp = self.authAPI('GET', '/api/v3/allOrders', { 'symbol': market })
+            
+            if isinstance(resp, list):
+                df = pd.DataFrame.from_dict(resp)
+            else: 
+                df = pd.DataFrame(resp, index=[0])
 
         if len(df) == 0:
             return pd.DataFrame()
 
-        # fix: float division by zero
-        # remove pending orders with status new (i.e. order made by user at limit)
-        df = df[df['status'] == "FILLED"]
+        # feature engineering
 
-        # replace null NaN values with 0
-        df.fillna(0, inplace=True)
+        def convert_time(epoch: int=0):
+            epoch_str = str(epoch)[0:10]
+            return datetime.fromtimestamp(int(epoch_str))
 
-        df = df[[ 'time', 'symbol', 'side', 'type', 'executedQty', 'cummulativeQuoteQty', 'status' ]]
-        df.columns = [ 'created_at', 'market', 'action', 'type', 'filled', 'size', 'status' ]
-        df['created_at'] = df['created_at'].apply(lambda x: int(str(x)[:10]))
-        df['created_at'] = df['created_at'].astype("datetime64[s]")
-        df['size'] = df['size'].astype(float)
-        df['filled'] = df['filled'].astype(float)
-        df['action'] = df['action'].str.lower()
-        df['type'] = df['type'].str.lower()
+        df.time = df['time'].map(convert_time)
+        df['time'] = pd.to_datetime(df['time']).dt.tz_localize('UTC')
+
+        df['size'] = np.where(df['side']=='BUY', df['cummulativeQuoteQty'], np.where(df['side']=='SELL', df['executedQty'], 222))
+        df['fees'] = df['size'].astype(float) * 0.001
+        df['fees'] = df['fees'].astype(object)
+
+        df['side'] = df['side'].str.lower() 
+
+        df.rename(columns={ 
+            'time': 'created_at', 
+            'symbol': 'market', 
+            'side': 'action', 
+            'executedQty': 'filled' 
+        }, errors='raise', inplace=True)
+
+        def convert_status(status: str=''):
+            if status == 'FILLED':
+                return 'done'
+            elif status == 'NEW':
+                return 'open'
+            elif status == 'PARTIALLY_FILLED':
+                return 'pending'
+            else:
+                return status    
+
+        df.status = df.status.map(convert_status)
         df['status'] = df['status'].str.lower()
-        df['price'] = df['size'] / df['filled']
 
-        # pylint: disable=unused-variable
-        for k, v in df.items():
-            if k == 'status':
-                df[k] = df[k].map(self.__convertStatus)
+        # select columns
+        df = df[[ 'created_at', 'market', 'action', 'type', 'size', 'filled', 'fees', 'price', 'status' ]]
 
+        # filtering
         if action != '':
             df = df[df['action'] == action]
-            df = df.reset_index(drop=True)
-
-        if status != 'all' and status != '':
+        if status != 'all':
             df = df[df['status'] == status]
-            df = df.reset_index(drop=True)
 
         return df
 
+    def getTime(self) -> datetime:
+        """Retrieves the exchange time"""
+    
+        def convert_time(epoch: int=0):
+            epoch_str = str(epoch)[0:10]
+            return datetime.fromtimestamp(int(epoch_str))
 
-    def marketBuy(self, market: str='', quote_quantity: float=0) -> list:
+        try:
+            # GET /api/v3/time
+            resp = self.authAPI('GET', '/api/v3/time')
+            return convert_time(int(resp['serverTime']))
+        except:
+            return None
+
+    def getMarketInfoFilters(self, market: str) -> pd.DataFrame:
+        """Retrieves markets exchange info"""
+
+        # GET /api/v3/allOrders
+        resp = self.authAPI('GET', '/api/v3/exchangeInfo', { 'symbol': market })
+        df = pd.DataFrame()
+
+        if 'symbols' in resp:
+            if isinstance(resp['symbols'], list):
+                if 'filters' in resp['symbols'][0]:
+                    df = pd.DataFrame.from_dict(resp['symbols'][0]['filters'])
+            else: 
+                if 'filers' in resp['symbols'][0]:   
+                    df = pd.DataFrame(resp['symbols'][0]['filters'], index=[0])
+
+        return df
+
+    def getTradeFee(self, market: str) -> float:
+        """Retrieves the trade fees"""
+
+        # GET /sapi/v1/asset/tradeFee
+        resp = self.authAPI('GET', '/sapi/v1/asset/tradeFee', { 'symbol': market })
+     
+        if len(resp) == 1 and 'takerCommission' in resp[0]:
+            return float(resp[0]['takerCommission'])
+        else:
+            return DEFAULT_TRADE_FEE_RATE
+
+    def getTicker(self, market: str=DEFAULT_MARKET) -> tuple:
+        """Retrieves the market ticker"""
+
+        # validates the market is syntactically correct
+        if not self._isMarketValid(market):
+            raise TypeError('Binance market required.')
+
+        # GET /api/v3/ticker/price
+        resp = self.authAPI('GET', '/api/v3/ticker/price', { 'symbol': market })
+
+        now = datetime.today().strftime('%Y-%m-%d %H:%M:%S')
+
+        if 'price' in resp:
+            return (str(self.getTime()), float(resp['price']))
+        else:
+            return (now, 0.0)
+
+    def marketBuy(self, market: str='', quote_quantity: float=0, test: bool=False) -> list:
         """Executes a market buy providing a funding amount"""
 
         # validates the market is syntactically correct
@@ -272,16 +418,29 @@ class AuthAPI(AuthAPIBase):
             # execute market buy
             stepper = 10.0 ** precision
             truncated = math.trunc(stepper * base_quantity) / stepper
-            Logger.info('Order quantity after rounding and fees: ' + str(truncated))
 
-            return self.client.order_market_buy(symbol=market, quantity=truncated)
+            order = { 
+                'symbol': market, 
+                'side': 'BUY', 
+                'type': 'MARKET', 
+                'quantity': truncated 
+            }
+
+            Logger.debug(order)
+
+            # POST /api/v3/order/test
+            if test is True:
+                resp = self.authAPI('POST', '/api/v3/order/test', order)
+            else:
+                resp = self.authAPI('POST', '/api/v3/order', order)
+
+            return resp
         except Exception as err:
             ts = datetime.now().strftime("%d-%m-%Y %H:%M:%S")
             Logger.error(ts + ' Binance ' + ' marketBuy ' + str(err))
-            return []       
+            return []  
 
-
-    def marketSell(self, market: str='', base_quantity: float=0) -> list:
+    def marketSell(self, market: str='', base_quantity: float=0, test: bool=False) -> list:
         """Executes a market sell providing a crypto amount"""
 
         # validates the market is syntactically correct
@@ -302,95 +461,164 @@ class AuthAPI(AuthAPIBase):
             # execute market sell
             stepper = 10.0 ** precision
             truncated = math.trunc(stepper * base_quantity) / stepper
-            Logger.info('Order quantity after rounding and fees: ' + str(truncated))
-            return self.client.order_market_sell(symbol=market, quantity=truncated)
+
+            order = { 
+                'symbol': market, 
+                'side': 'SELL', 
+                'type': 'MARKET', 
+                'quantity': truncated 
+            }
+
+            Logger.debug(order)
+
+            # POST /api/v3/order/test
+            if test is True:
+                resp = self.authAPI('POST', '/api/v3/order/test', order)
+            else:
+                resp = self.authAPI('POST', '/api/v3/order', order)
+
+            return resp
         except Exception as err:
             ts = datetime.now().strftime("%d-%m-%Y %H:%M:%S")
             Logger.error(ts + ' Binance ' + ' marketSell ' + str(err))
             return []
 
+    def authAPI(self, method: str, uri: str, payload: str={}) -> dict:
+        """Initiates a REST API call to the exchange"""
 
-    def getTradeFee(self, market: str) -> float:
-        resp = self.client.get_trade_fee(symbol=market, timestamp=self.getTime())
-      
-        if 'tradeFee' not in resp:
-            Logger.info('*** getTradeFee(' + market + ') - missing "tradeFee" ***')
-            Logger.info(resp)
+        if not isinstance(method, str):
+            raise TypeError('Method is not a string.')
+
+        if not method in ['GET', 'POST']:
+            raise TypeError('Method not GET or POST.')
+
+        if not isinstance(uri, str):
+            raise TypeError('URI is not a string.')
+
+        signed_uri = [
+            '/api/v3/account',
+            '/api/v3/allOrders',
+            '/api/v3/order',
+            '/api/v3/order/test',
+            '/sapi/v1/asset/tradeFee'
+        ]
+
+        query_string = urlencode(payload, True)
+        if uri in signed_uri and query_string:
+            query_string = "{}&timestamp={}".format(query_string, self.getTimestamp())
+        elif uri in signed_uri:
+            query_string = 'timestamp={}'.format(self.getTimestamp())
+
+        if uri in signed_uri:
+            url = self._api_url + uri + '?' + query_string + '&signature=' + self.createHash(query_string)
         else:
-            if len(resp['tradeFee']) == 0:
-                Logger.info('*** getTradeFee(' + market + ') - "tradeFee" empty ***') 
-                Logger.info(resp)
+            url = self._api_url + uri + '?' + query_string
+        
+        params = {'url': url, 'params': {}}
+
+        try:
+            resp = self._dispatch_request(method)(**params)
+
+            if resp.status_code != 200:               
+                resp_message = resp.json()['msg']
+                message = f'{method} ({resp.status_code}) {self._api_url}{uri} - {resp_message}'
+                if self.die_on_api_error:
+                    raise Exception(message)
+                else:
+                    Logger.error(f'Error: {message}')
+                    return {}
+
+            resp.raise_for_status()
+            return resp.json()
+        
+        except requests.ConnectionError as err:
+            return self.handle_api_error(err, 'ConnectionError')
+
+        except requests.exceptions.HTTPError as err:
+            return self.handle_api_error(err, 'HTTPError')
+
+        except requests.Timeout as err:
+            return self.handle_api_error(err, 'Timeout')
+
+        except json.decoder.JSONDecodeError as err:
+            return self.handle_api_error(err, 'JSONDecodeError')
+
+    def handle_api_error(self, err: str, reason: str) -> dict:
+        """Handler for API errors"""
+
+        if self.debug:
+            if self.die_on_api_error:
+                raise SystemExit(err)
             else:
-                if 'taker' not in resp['tradeFee'][0]:
-                    Logger.info('*** getTradeFee(' + market + ') - missing "trader" ***')
-                    Logger.info(resp)                    
-
-        if 'success' in resp:
-            return resp['tradeFee'][0]['taker']
+                Logger.debug(err)
+                return {}
         else:
-            return DEFAULT_TRADE_FEE_RATE
+            if self.die_on_api_error:
+                raise SystemExit(f"{reason}: {self._api_url}")
+            else:
+                Logger.info(f"{reason}: {self._api_url}")
+                return {}
 
+class PublicAPI(AuthAPIBase):
+    def __init__(self, api_url='https://api.binance.com') -> None:
+        """Binance API object model
+    
+        Parameters
+        ----------
+        api_url
+            Binance API URL
+        """
 
-    def getMarketInfo(self, market: str) -> dict:
-        # validates the market is syntactically correct
-        if not self._isMarketValid(market):
-            raise TypeError('Binance market required.')
+        # options
+        self.debug = False
+        self.die_on_api_error = False
 
-        return self.client.get_symbol_info(symbol=market)
+        valid_urls = [
+            'https://api.binance.com',
+            'https://api.binance.us',
+            'https://testnet.binance.vision'
+        ]
 
+        # validate Binance API
+        if api_url not in valid_urls:
+            raise ValueError('Binance API URL is invalid')
 
-    def getMarketInfoFilters(self, market: str) -> pd.DataFrame:
-        return pd.DataFrame(self.client.get_symbol_info(symbol=market)['filters'])
-
-
-    def getTicker(self, market:str) -> tuple:
-        # validates the market is syntactically correct
-        if not self._isMarketValid(market):
-            raise TypeError('Binance market required.')
-
-        resp = self.client.get_symbol_ticker(symbol=market)
-
-        if 'price' in resp:
-            return (self.getTime().strftime('%Y-%m-%d %H:%M:%S'), float('{:.8f}'.format(float(resp['price']))))
-
-        now = datetime.today().strftime('%Y-%m-%d %H:%M:%S')
-        return (now, 0.0)
-
+        self._api_url = api_url
 
     def getTime(self) -> datetime:
         """Retrieves the exchange time"""
     
+        def convert_time(epoch: int=0):
+            epoch_str = str(epoch)[0:10]
+            return datetime.fromtimestamp(int(epoch_str))
+
         try:
-            resp = self.client.get_server_time()
-            epoch = int(str(resp['serverTime'])[0:10])
-            return datetime.fromtimestamp(epoch)
+            # GET /api/v3/time
+            resp = self.authAPI('GET', '/api/v3/time')
+            return convert_time(int(resp['serverTime']))
         except:
-            return datetime.now()
+            return None
 
+    def getTicker(self, market: str=DEFAULT_MARKET) -> tuple:
+        """Retrives the market ticker"""
+        
+        # validates the market is syntactically correct
+        if not self._isMarketValid(market):
+            raise TypeError('Binance market required.')
 
-class PublicAPI(AuthAPIBase):
-    def __init__(self) -> None:
-        for i in range(10):
-            try:
-                self.client = Client()
-                break
-            except Exception as e:
-                if i == 9:
-                    raise SystemExit("Can not create instance of AuthAPI client.")
-                Logger.error('Exception: ' + str(e)) 
-                Logger.error('Error on creating instance of AuthAPI Client. Trying again... Attempt: ' + str(i))
-                sleep(0.1)
-               
+        # GET /api/v3/ticker/price
+        resp = self.authAPI('GET', '/api/v3/ticker/price', { 'symbol': market })
 
-    def __truncate(self, f, n) -> int:
-        return math.floor(f * 10 ** n) / 10 ** n
+        now = datetime.today().strftime('%Y-%m-%d %H:%M:%S')
 
-    
-    def getClient(self) -> Client:
-        return self.client
+        if 'price' in resp:
+            return (str(self.getTime()), float(resp['price']))
+        else:
+            return (now, 0.0)
 
-    
     def getHistoricalData(self, market: str=DEFAULT_MARKET, granularity: str=DEFAULT_GRANULARITY, iso8601start: str='', iso8601end: str='') -> pd.DataFrame:
+        """Retrieves historical market data"""
+
         # validates the market is syntactically correct
         if not self._isMarketValid(market):
             raise TypeError('Binance market required.')
@@ -403,53 +631,27 @@ class PublicAPI(AuthAPIBase):
         if not granularity in SUPPORTED_GRANULARITY:
             raise TypeError('Granularity options: ' + ", ".join(map(str, SUPPORTED_GRANULARITY)))
 
-        # validates the ISO 8601 start date is a string (if provided)
-        if not isinstance(iso8601start, str):
-            raise TypeError('ISO8601 start integer as string required.')
-
         # validates the ISO 8601 end date is a string (if provided)
         if not isinstance(iso8601end, str):
             raise TypeError('ISO8601 end integer as string required.')
 
-        # if only a start date is provided
         if iso8601start != '' and iso8601end == '':
-            try:
-                multiplier = MULTIPLIER_EQUIVALENTS[SUPPORTED_GRANULARITY.index(granularity)] 
-            except:
-                multiplier = 1
-    
-            # calculate the end date using the granularity
-            iso8601end = str((datetime.strptime(iso8601start, '%Y-%m-%dT%H:%M:%S.%f') + timedelta(minutes=granularity * multiplier)).isoformat())
-
-        if iso8601start != '' and iso8601end != '':
-            Logger.info('Attempting to retrieve data from ' + iso8601start)
-            resp = self.client.get_historical_klines(market, granularity, iso8601start, iso8601end)
-
-            if len(resp) > 300:
-                resp = resp[:300]
-        else: # TODO: replace with a KLINE_MESSAGE_FOO equivalent
-            if granularity == '1m':
-                resp = self.client.get_historical_klines(market, granularity, '12 hours ago UTC')
-                resp = resp[-300:]
-            elif granularity == '5m':
-                resp = self.client.get_historical_klines(market, granularity, '2 days ago UTC')
-                resp = resp[-300:]
-            elif granularity == '15m':
-                resp = self.client.get_historical_klines(market, granularity, '4 days ago UTC')
-                resp = resp[-300:]
-            elif granularity == '1h':
-                resp = self.client.get_historical_klines(market, granularity, '13 days ago UTC')
-                resp = resp[-300:]
-            elif granularity == '6h':
-                resp = self.client.get_historical_klines(market, granularity, '75 days ago UTC')
-                resp = resp[-300:]
-            elif granularity == '1d':
-                resp = self.client.get_historical_klines(market, granularity, '251 days ago UTC')
-            else:
-                raise Exception('Something went wrong!')
-                   
+            startTime = int((datetime.strptime(iso8601start, '%Y-%m-%dT%H:%M:%S') - datetime(1970, 1, 1)).total_seconds())
+           
+            # GET /api/v3/klines
+            resp = self.authAPI('GET', '/api/v3/klines', { 'symbol': market, 'interval': granularity, 'startTime': startTime, 'limit': 300 })  
+        elif iso8601start != '' and iso8601end != '':
+            startTime = int((datetime.strptime(iso8601start, '%Y-%m-%dT%H:%M:%S') - datetime(1970, 1, 1)).total_seconds())
+           
+            # GET /api/v3/klines
+            resp = self.authAPI('GET', '/api/v3/klines', { 'symbol': market, 'interval': granularity, 'startTime': startTime, 'limit': 300 })
+        else:
+            # GET /api/v3/klines
+            resp = self.authAPI('GET', '/api/v3/klines', { 'symbol': market, 'interval': granularity, 'limit': 300 })
+        
         # convert the API response into a Pandas DataFrame
         df = pd.DataFrame(resp, columns=[ 'open_time', 'open', 'high', 'low', 'close', 'volume', 'close_time', 'quote_asset_volume', 'number_of_trades', 'taker_buy_base_asset_volume', 'traker_buy_quote_asset_volume', 'ignore' ])
+
         df['market'] = market
         df['granularity'] = granularity
 
@@ -477,6 +679,10 @@ class PublicAPI(AuthAPIBase):
             df.index.names = ['ts']           
             df['date'] = tsidx
 
+        # if specified, fix end time
+        if iso8601end != '':
+            df = df[df['date'] <= iso8601end]
+
         # re-order columns
         df = df[[ 'date', 'market', 'granularity', 'low', 'high', 'open', 'close', 'volume' ]]
 
@@ -492,28 +698,57 @@ class PublicAPI(AuthAPIBase):
 
         return df
 
-    
-    def getTicker(self, market: str) -> tuple:
-        # validates the market is syntactically correct
-        if not self._isMarketValid(market):
-            raise TypeError('Binance market required.')
+    def authAPI(self, method: str, uri: str, payload: str={}) -> dict:
+        """Initiates a REST API call to exchange"""
+        
+        if not isinstance(method, str):
+            raise TypeError('Method is not a string.')
 
-        resp = self.client.get_symbol_ticker(symbol=market)
+        if not method in ['GET', 'POST']:
+            raise TypeError('Method not GET or POST.')
 
-        if 'price' in resp:
-            return (self.getTime().strftime('%Y-%m-%d %H:%M:%S'), float('{:.8f}'.format(float(resp['price']))))
+        if not isinstance(uri, str):
+            raise TypeError('URI is not a string.')
 
-        now = datetime.today().strftime('%Y-%m-%d %H:%M:%S')
-        return (now, 0.0)
-
-    
-    def getTime(self) -> datetime:
-        """Retrieves the exchange time"""
-    
         try:
-            resp = self.client.get_server_time()
-            epoch = int(str(resp['serverTime'])[0:10])
-            return datetime.fromtimestamp(epoch)
-        except:
-            return datetime.now()
+            resp = requests.get(f'{self._api_url}{uri}', params=payload)
 
+            if resp.status_code != 200:
+                resp_message = resp.json()['msg']
+                message = f'{method} ({resp.status_code}) {self._api_url}{uri} - {resp_message}'
+                if self.die_on_api_error:
+                    raise Exception(message)
+                else:
+                    Logger.error(f'Error: {message}')
+                    return {}
+
+            resp.raise_for_status()
+            return resp.json()
+        
+        except requests.ConnectionError as err:
+            return self.handle_api_error(err, 'ConnectionError')
+
+        except requests.exceptions.HTTPError as err:
+            return self.handle_api_error(err, 'HTTPError')
+
+        except requests.Timeout as err:
+            return self.handle_api_error(err, 'Timeout')
+
+        except json.decoder.JSONDecodeError as err:
+            return self.handle_api_error(err, 'JSONDecodeError')
+
+    def handle_api_error(self, err: str, reason: str) -> dict:
+        """Handler for API errors"""
+
+        if self.debug:
+            if self.die_on_api_error:
+                raise SystemExit(err)
+            else:
+                Logger.debug(err)
+                return {}
+        else:
+            if self.die_on_api_error:
+                raise SystemExit(f"{reason}: {self._api_url}")
+            else:
+                Logger.info(f"{reason}: {self._api_url}")
+                return {}
